requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
np==1.0.2
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
plotly==5.13.1
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.26.0.post1
tabulate==0.9.0
tenacity==8.2.2
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
